hopr_transport_probe

Module ping

Structs

PingConfig

Configuration for the [Ping] mechanism

PingQueryReplier

Helper object allowing to send a ping query as a wrapped channel combination that can be filled up on the transport part and awaited locally by the Pinger.

Pinger

Implementation of the ping mechanism

Type Aliases

HeartbeatSendPingTx

Heartbeat send ping TX type

PingQueryResult

Ping query result type holding data about the ping duration and the string containg an optional version information of the pinged peer, if provided.